Optimized Server Rest API

+++ client/src/components/register.component.js
@@ -21,7 +21,7 @@ const Register = () => {
         const btn = document.getElementById('register');
         async function submitData(){
             btn.innerHTML = "Registering...";
-            const registerData = { email, password }
+            const registerData = { email, password, confirmPassword: confirmPsw }
             await axios.post(`${SERVER_URL}/account/register`, registerData, { headers: { 'X-CSRF-TOKEN': getCSRFToken()[0], 'X-XSRF-TOKEN': getCSRFToken()[1] } })
             .then(res => {
                 localStorage.setItem('__id', res.data.id);
